give reward just the first time you talk with the first npc
I found useless to show the reward multiple times if we are sure that the player doesn't have a wallet.
On the other hand if the player does, and it click dismiss, every time that it will talk with the first npc the reward will appear until the player collects it.